Safe_backend/requirements.txt
flask
datetime
geopy
google-maps-routeoptimization
gogole-generativeai==0.8.3
gunicorn
apscheduler
requests
